+++ src/py/py15.estimate_sequencing_depth.py
@@ -1,7 +1,7 @@
 import sys
 target_bed = open(sys.argv[1])
 base_mapped = open(sys.argv[2])
-n_read_base = float(base_mapped.readline().strip())
+n_read_base = double(base_mapped.readline().strip())
 n_region_base = 0
 for line in target_bed:
 	sp = line.strip().split('\t')
@@ -10,6 +10,6 @@ for line in target_bed:
 base_mapped.close()
 target_bed.close()
 output = open(sys.argv[3], 'w')
-output.write(str(float(n_read_base)/float(n_region_base) * 0.8) + '\n')
+output.write(str(double(n_read_base)/double(n_region_base) * 0.8) + '\n')
 output.close()
